Code review comments
