+++ src/Core.UnitTests/CFamily/EffectiveRulesConfigCalculatorTests.cs
@@ -21,7 +21,6 @@
 using System.Collections.Generic;
 using FluentAssertions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Moq;
 using SonarLint.VisualStudio.Core.CFamily;
 using SonarLint.VisualStudio.Integration.UnitTests;
 using SonarLint.VisualStudio.Integration.UnitTests.CFamily;
@@ -31,15 +30,22 @@ namespace SonarLint.VisualStudio.Core.UnitTests.CFamily
     [TestClass]
     public class EffectiveRulesConfigCalculatorTests
     {
+        private TestLogger testLogger;
+        private EffectiveRulesConfigCalculator testSubject;
+
+        [TestInitialize]
+        public void TestInitialize()
+        {
+            testLogger = new TestLogger();
+            testSubject = new EffectiveRulesConfigCalculator(testLogger);
+        }
+
         [TestMethod]
         [DataRow(true /* custom settings are null */)]
         [DataRow(false /* custom settings are not null, but emtpy */)]
-        public void GetConfig_NoCustomSettings_NoError_DefaultsReturned(bool customSettingsAreNull)
+        public void GetConfig_NoCustomSettings_DefaultsReturned(bool customSettingsAreNull)
         {
             // Arrange
-            var testLogger = new TestLogger();
-            var testSubject = new EffectiveRulesConfigCalculator(testLogger);
-
             var defaultRulesConfig = CreateWellKnownRulesConfig("language1");
             var sourcesSettings = customSettingsAreNull ? null : new RulesSettings();
 
@@ -47,7 +53,6 @@ namespace SonarLint.VisualStudio.Core.UnitTests.CFamily
             var result = testSubject.GetEffectiveRulesConfig("language1", defaultRulesConfig, sourcesSettings);
 
             // Assert
-            result.Should().NotBeSameAs(defaultRulesConfig);
             result.LanguageKey.Should().Be("language1");
             result.AllPartialRuleKeys.Should().BeEquivalentTo(defaultRulesConfig.AllPartialRuleKeys);
 
@@ -58,9 +63,6 @@ namespace SonarLint.VisualStudio.Core.UnitTests.CFamily
         public void GetConfig_RulesInCustomSettings_MergedConfigReturned()
         {
             // Arrange
-            var testLogger = new TestLogger();
-            var testSubject = new EffectiveRulesConfigCalculator(testLogger);
-
             var defaultRulesConfig = CreateWellKnownRulesConfig("key");
             var sourcesSettings = new RulesSettings
             {
@@ -68,12 +70,12 @@ namespace SonarLint.VisualStudio.Core.UnitTests.CFamily
                 {
                     // Turn an active rule off...
                     { "key:" + WellKnownPartialRuleKey1_Active, new RuleConfig { Level = RuleLevel.Off } },
-                    // ... and an inactive rule one
+                    // ... and an inactive rule on
                     { "key:" + WellKnownPartialRuleKey3_Inactive, new RuleConfig { Level = RuleLevel.On } }
                 }
             };
 
-            var result = testSubject.GetEffectiveRulesConfig("language1", defaultRulesConfig, sourcesSettings);
+            var result = testSubject.GetEffectiveRulesConfig("key", defaultRulesConfig, sourcesSettings);
 
             result.LanguageKey.Should().Be("key");
             result.AllPartialRuleKeys.Should().BeEquivalentTo(defaultRulesConfig.AllPartialRuleKeys);
@@ -84,13 +86,10 @@ namespace SonarLint.VisualStudio.Core.UnitTests.CFamily
         public void GetConfig_CachedResultsReturnedIfAvailable()
         {
             // Arrange
-            var testLogger = new TestLogger();
-            var testSubject = new EffectiveRulesConfigCalculator(testLogger);
-
             var defaultRulesConfig = CreateWellKnownRulesConfig("key");
             var sourcesSettings = new RulesSettings
             {
-                Rules = new System.Collections.Generic.Dictionary<string, RuleConfig>
+                Rules = new Dictionary<string, RuleConfig>
                 {
                     { "rule1", new RuleConfig() }
                 }
@@ -99,6 +98,7 @@ namespace SonarLint.VisualStudio.Core.UnitTests.CFamily
             // 1. First call -> new config returned
             var result1 = testSubject.GetEffectiveRulesConfig("language1", defaultRulesConfig, sourcesSettings);
 
+            result1.Should().NotBeNull();
             result1.Should().NotBeSameAs(defaultRulesConfig);
             testLogger.AssertOutputStringExists(CoreStrings.EffectiveRules_CacheMiss);
 
@@ -116,85 +116,6 @@ namespace SonarLint.VisualStudio.Core.UnitTests.CFamily
             result3.Should().NotBeSameAs(result2);
             testLogger.AssertOutputStringExists(CoreStrings.EffectiveRules_CacheMiss);
         }
-
-        #region Cache tests
-
-        [TestMethod]
-        public void Cache_DifferentSourceConfig_NotFound_AndEntryCleared()
-        {
-            var sourceConfig1 = new Mock<ICFamilyRulesConfig>().Object;
-            var sourceSettings1 = new RulesSettings();
-            var effectiveConfig1 = new Mock<ICFamilyRulesConfig>().Object;
-
-            var testSubject = new EffectiveRulesConfigCalculator.RulesConfigCache();
-
-            testSubject.Add("key1", sourceConfig1, sourceSettings1, effectiveConfig1);
-            testSubject.CacheCount.Should().Be(1);
-
-            // 1. Search for added item -> found
-            testSubject.FindConfig("key1", sourceConfig1, sourceSettings1).Should().BeSameAs(effectiveConfig1);
-            testSubject.CacheCount.Should().Be(1);
-
-            // 2. Different source config -> not found
-            testSubject.FindConfig("key1", new Mock<ICFamilyRulesConfig>().Object, sourceSettings1).Should().BeNull();
-            testSubject.CacheCount.Should().Be(0);
-        }
-
-        [TestMethod]
-        public void Cache_DifferentSourceSettings_NotFound_AndEntryCleared()
-        {
-            var sourceConfig1 = new Mock<ICFamilyRulesConfig>().Object;
-            var sourceSettings1 = new RulesSettings();
-            var effectiveConfig1 = new Mock<ICFamilyRulesConfig>().Object;
-
-            var testSubject = new EffectiveRulesConfigCalculator.RulesConfigCache();
-
-            testSubject.Add("key1", sourceConfig1, sourceSettings1, effectiveConfig1);
-            testSubject.CacheCount.Should().Be(1);
-
-            // 1. Search for added item -> found
-            testSubject.FindConfig("key1", sourceConfig1, sourceSettings1).Should().BeSameAs(effectiveConfig1);
-            testSubject.CacheCount.Should().Be(1);
-
-            // 2. Different source settings -> not found
-            testSubject.FindConfig("key1", sourceConfig1, new RulesSettings()).Should().BeNull();
-            testSubject.CacheCount.Should().Be(0);
-        }
-
-        [TestMethod]
-        public void Cache_MultipleEntries()
-        {
-            var sourceConfig1 = new Mock<ICFamilyRulesConfig>().Object;
-            var sourceConfig2 = new Mock<ICFamilyRulesConfig>().Object;
-
-            var sourceSettings1 = new RulesSettings();
-            var sourceSettings2 = new RulesSettings();
-
-            var effectiveConfig1 = new Mock<ICFamilyRulesConfig>().Object;
-            var effectiveConfig2 = new Mock<ICFamilyRulesConfig>().Object;
-
-            var testSubject = new EffectiveRulesConfigCalculator.RulesConfigCache();
-
-            // 1. Empty cache -> cache miss
-            testSubject.FindConfig("key1", sourceConfig1, sourceSettings1).Should().BeNull();
-
-            // 2. Add first entry to cache
-            testSubject.Add("key1", sourceConfig1, sourceSettings1, effectiveConfig1);
-            testSubject.CacheCount.Should().Be(1);
-
-            // 3. Find second language - not found
-            testSubject.FindConfig("key2", sourceConfig2, sourceSettings2).Should().BeNull();
-
-            // 4. Add second entry to cache
-            testSubject.Add("key2", sourceConfig2, sourceSettings2, effectiveConfig2);
-            testSubject.CacheCount.Should().Be(2);
-
-            // 5. Check can find both entries
-            testSubject.FindConfig("key1", sourceConfig1, sourceSettings1).Should().BeSameAs(effectiveConfig1);
-            testSubject.FindConfig("key2", sourceConfig2, sourceSettings2).Should().BeSameAs(effectiveConfig2);
-        }
-
-        #endregion Cache tests
 
         internal const string WellKnownPartialRuleKey1_Active = "rule1";
         internal const string WellKnownPartialRuleKey2_Active = "rule2";
